common

+++ image_processing/preprocess_images.py
@@ -295,7 +295,7 @@ def preprocess_image(tiff_file, outputdir='.', invert=None, bg_subtract=True, bg
                 image = images[c][r]
                 i0 = np.min(image)
                 i1 = np.max(image)
-                image = (np.array(image) - i0)/(i1-i0)
+                image = (np.array(image, dtype=np.float_) - i0)/(i1-i0)
                 ax=fig.add_subplot(gs[r,c])
                 axes.append(ax)
                 cf=ax.imshow(image, cmap=cmaps[c])
